Cleanup
All of the scripts and declared values were missing from the prefabs, the card database, and the game scene. Everything is fixed and tested.

The scaling on scenes are fixed.

Couldn't fix the Shuffle function as I lack the brain capacity to do so. Instead, drawn cards are randomized.

Added Networking to game scene:
 - Added drawing cards with authority
 - Added Drag and Drop functionalities
 - Added Play Area functionality